Scenarios run bugs fixed. Add store/read main scenario cmds.
Added command for storing and reading main scenario. Deserialize now
parses all until new line character.

Added reaction on error during sending over CAN bus. Crc8 calcs extended
to calc up to 65535 bytes data length. To CrcPartial added compile-time
checks. UpdateCrc and CheckCrc moved to separate functions.

Tests improved, two more tests added (when error happens on sending over
CAN bus). Also logs now works in tests also, with printing to stdout.

+++ Arduino/libraries/CRC/crc.cpp
@@ -17,7 +17,7 @@ static const uint8_t PROGMEM dscrc2x16_table[] = {
 
 // Compute a Dallas Semiconductor 8 bit CRC. These show up in the ROM
 // and the registers.  (Use tiny 2x16 entry CRC table)
-uint8_t crc8(const uint8_t *addr, uint8_t len)
+uint8_t crc8(const uint8_t *addr, uint16_t len)
 {
         uint8_t crc = 0;
 
@@ -34,7 +34,7 @@ uint8_t crc8(const uint8_t *addr, uint8_t len)
 // Compute a Dallas Semiconductor 8 bit CRC directly.
 // this is much slower, but a little smaller, than the lookup table.
 //
-uint8_t crc8(const uint8_t *addr, uint8_t len)
+uint8_t crc8(const uint8_t *addr, uint16_t len)
 {
         uint8_t crc = 0;
