+++ Arduino/libraries/CRC/crc.h
@@ -10,7 +10,7 @@
 
 // Compute a Dallas Semiconductor 8 bit CRC, these are used in the
 // ROM and scratchpad registers.
-uint8_t crc8(const uint8_t *addr, uint8_t len);
+uint8_t crc8(const uint8_t *addr, uint16_t len);
 
 // Compute the 1-Wire CRC16 and compare it against the received CRC.
 // Example usage (reading a DS2408):
